Silicosis_pg - proteinGroups data of a silicosis mouse model

The Silicosis_pg is proteinGroups result of a proteomics study on silica-exposed mouse model reported by Wang, Zhang, Liu, et al.

Usage

Silicosis_pg

Format

A data.frame with 10718 rows and 55 variables:

Protein.IDs

Uniprot IDs

Majority.protein.IDs

Uniprot IDs of major protein(s) in the protein group

Peptide.counts..all.

Total peptides counts of each protein

Peptide.counts..razor.unique.

Razor or unique peptides counts of each protein

Peptide.counts..unique.

Unique peptides counts of each protein

Protein.names

Full protein names

Gene.names

Gene name

Fasta.headers

Header as present in the Uniprot fasta file

Peptides

Number of peptides identified for this protein group

Razor...unique.peptides

Number of peptides used for the quantification of this protein group

Unique.peptides

Number of peptides identified which are unique for this protein group

iBAQ columns

intensity-based absolute-protein-quantification

LFQ.intensity columns

LFQ normalized mass spectrometry intensity, A.U.

Only.identified.by.site

The protein is only identified by a modification site if marked ('+')

Reverse

The protein is identified in the decoy database if marked ('+')

Potential.contaminant

The protein is a known contaminant if marked ('+')

Source

Wang M, Zhang Z, Liu J, et al. Gefitinib and fostamatinib target EGFR and SYK to attenuate silicosis: a multi-omics study with drug exploration. Signal Transduct Target Ther. 2022;7(1):157. doi: 10.1038/s41392-022-00959-3.