Generate a results table

get_results generates a results table from a proteomics or RNA-seq dataset on which differential analysis was performed.

Usage

get_results(object)

Arguments

object

SummarizedExperiment or DEGdata, (output from test_diff() and add_rejections()).

Value

A data.frame object containing all results variables from the performed analysis.

Examples

#Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Differential test
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
#> filter base on missing number is <= 0 in at least one condition.
#> filter base on missing number fraction < 0.4 in each row
#> filter base on giving formula 
norm <- normalize_vsn(filt)
#> vsn2: 8762 x 20 matrix (1 stratum). 
#> Please use 'meanSdPlot' to verify the fit.
imputed <- impute(norm, fun = "MinProb", q = 0.05)
#> [1] 0.3026531
diff <- test_diff(imputed, type = "control", control  = c("PBS"), fdr.type = "Storey's qvalue")
#> Tested contrasts: W10_vs_PBS, W2_vs_PBS, W4_vs_PBS, W6_vs_PBS, W9_vs_PBS
#> Storey's qvalue
dep <- add_rejections(diff, alpha = 0.01,lfc = 2)

# Get results
results <- get_results(dep)
colnames(results)
#>  [1] "name"                   "ID"                     "W10_vs_PBS_p.val"      
#>  [4] "W2_vs_PBS_p.val"        "W4_vs_PBS_p.val"        "W6_vs_PBS_p.val"       
#>  [7] "W9_vs_PBS_p.val"        "W10_vs_PBS_p.adj"       "W2_vs_PBS_p.adj"       
#> [10] "W4_vs_PBS_p.adj"        "W6_vs_PBS_p.adj"        "W9_vs_PBS_p.adj"       
#> [13] "W10_vs_PBS_significant" "W2_vs_PBS_significant"  "W4_vs_PBS_significant" 
#> [16] "W6_vs_PBS_significant"  "W9_vs_PBS_significant"  "significant"           
#> [19] "W10_vs_PBS_ratio"       "W2_vs_PBS_ratio"        "W4_vs_PBS_ratio"       
#> [22] "W6_vs_PBS_ratio"        "W9_vs_PBS_ratio"        "PBS_centered"          
#> [25] "W10_centered"           "W2_centered"            "W4_centered"           
#> [28] "W6_centered"            "W9_centered"           

significant_proteins <- results[results$significant,]
head(significant_proteins)
#>      name     ID W10_vs_PBS_p.val W2_vs_PBS_p.val W4_vs_PBS_p.val
#> 1    Acp5 Q05117     1.411753e-05    5.910320e-04    4.778377e-03
#> 2 Adamts8 P57110     2.309899e-01    4.039742e-02    2.257890e-02
#> 3    Add2 Q9QYB8     2.040945e-03    6.075769e-03    2.625233e-04
#> 4    Aif1 O70200     5.802446e-06    7.746131e-08    1.653705e-07
#> 5  Alox12 P39655     3.538216e-03    3.931429e-03    4.797644e-03
#> 6    Ank1 Q02357     2.101053e-04    4.827071e-04    1.665076e-05
#>   W6_vs_PBS_p.val W9_vs_PBS_p.val W10_vs_PBS_p.adj W2_vs_PBS_p.adj
#> 1    2.099730e-04    3.195635e-04         0.000719        2.44e-03
#> 2    8.781493e-04    9.696470e-03         0.194000        4.54e-02
#> 3    2.525880e-03    9.264588e-03         0.009990        1.18e-02
#> 4    5.411303e-06    3.620109e-04         0.000479        1.09e-05
#> 5    1.476969e-04    5.409628e-03         0.013400        8.67e-03
#> 6    7.312273e-05    7.952478e-05         0.002970        2.12e-03
#>   W4_vs_PBS_p.adj W6_vs_PBS_p.adj W9_vs_PBS_p.adj W10_vs_PBS_significant
#> 1        1.32e-02        0.001170         0.00290                   TRUE
#> 2        3.63e-02        0.002820         0.02260                  FALSE
#> 3        2.26e-03        0.005810         0.02190                   TRUE
#> 4        3.51e-05        0.000140         0.00310                   TRUE
#> 5        1.32e-02        0.000920         0.01530                  FALSE
#> 6        4.78e-04        0.000607         0.00138                   TRUE
#>   W2_vs_PBS_significant W4_vs_PBS_significant W6_vs_PBS_significant
#> 1                 FALSE                 FALSE                 FALSE
#> 2                 FALSE                 FALSE                  TRUE
#> 3                 FALSE                  TRUE                 FALSE
#> 4                  TRUE                  TRUE                  TRUE
#> 5                 FALSE                 FALSE                  TRUE
#> 6                 FALSE                  TRUE                  TRUE
#>   W9_vs_PBS_significant significant W10_vs_PBS_ratio W2_vs_PBS_ratio
#> 1                 FALSE        TRUE             2.26            1.29
#> 2                 FALSE        TRUE             1.23            1.79
#> 3                 FALSE        TRUE            -2.28           -1.60
#> 4                 FALSE        TRUE             2.74            3.10
#> 5                 FALSE        TRUE            -1.80           -1.45
#> 6                  TRUE        TRUE            -2.35           -1.76
#>   W4_vs_PBS_ratio W6_vs_PBS_ratio W9_vs_PBS_ratio PBS_centered W10_centered
#> 1           0.994            1.44            1.69        -1.14        1.120
#> 2           2.030            3.26            2.89        -1.83       -0.599
#> 3          -2.350           -1.81           -1.84         1.57       -0.716
#> 4           2.940            2.25            1.87        -2.12        0.618
#> 5          -1.410           -2.12           -1.70         1.35       -0.457
#> 6          -2.420           -2.12           -2.58         1.75       -0.593
#>   W2_centered W4_centered W6_centered W9_centered
#> 1     0.15100     -0.1470       0.300       0.549
#> 2    -0.03470      0.2000       1.430       1.060
#> 3    -0.03730     -0.7880      -0.247      -0.273
#> 4     0.98400      0.8190       0.129      -0.247
#> 5    -0.10500     -0.0643      -0.773      -0.351
#> 6    -0.00249     -0.6710      -0.370      -0.826