Make unique names for a modication-enriched peptide table

make_unique_ptm generates unique identifiers for a proteomics dataset based on "modified_name" columns. If "modified_name" is absent, it will creat a modified name based on "gene_name"/"Protein", "aa" and "pos".

Usage

make_unique_ptm(
  PTMdata,
  gene_name = "Gene.names",
  protein_ID = "Protein",
  aa = "Amino.acid",
  pos = "Position",
  modified_name = NULL,
  delim = ";"
)

Arguments

PTMdata

Data.frame, modified peptide table for which unique names and ID for each modified peptide will be created.

gene_name

Character(1), Name of the column containing gene symbol or protein name.

protein_ID

Character(1), Name of the column containing protein ID or gene ID.

aa

Character(1), Name of the column containing abbreviation of modified amino acid, e.g. "K", "S", "T". It is unnecessary if modified_name is provided.

pos

Character(1), Name of the column containing position of modification on proteins. It is unnecessary if modified_name is provided.

modified_name

Character(1), Name of the column containing names of modified peptide in the format (gene name)_(modified amino acid)(position of modification), e.g. "TBCA_K51".

delim

Character(1), Sets the delimiter separating the multiple feature names within one protein group(gene_name and protein_ID).

Value

A data.frame with the 6 additional variables "name" and "ID" containing unique names and identifiers or modified peptide respectively, e.g. "TBCA_K51" and "O75347_K51". And "gene_name", protein_ID" save protein information of modified peptide, e.g. "TBCA" and "O75347". "modified_aa" and "modified_pos" is the modified amino acid and position, e.g. "K" and "51".

Details

If modified_name is provide, gene_name, protein_ID, aa and pos are unnessary. Additional variables is generated by parsing modified_name. else gene_name or protein_ID, aa and pos are nessary.

Examples

## phosphorylated peptides table of the silicosis mouse model.
data(Silicosis_phos)
## Format the modification information and generated modified-peptides identifier.
unique_pho <- make_unique_ptm(Silicosis_phos, gene_name = "Gene.names",
                              protein_ID = "Protein", aa = "Amino.acid",
                              pos = "Position")
head(unique_pho[, c("name","ID")], 10)
#>            name          ID
#> 1    P0DPB4_S60  P0DPB4_S60
#> 2    P0DPB4_S61  P0DPB4_S61
#> 3    P0DPB4_S62  P0DPB4_S62
#> 4    P0DPB4_S63  P0DPB4_S63
#> 5    P0DPB4_S67  P0DPB4_S67
#> 6  Ccdc173_S151 A0JLY1_S151
#> 7  Tbc1d25_S560 A1A5B6_S560
#> 8  Tbc1d25_S211 A1A5B6_S211
#> 9      Fscb_S32  A1EGX6_S32
#> 10   Fam65c_S24  A1L3T7_S24