Plot values for a protein/gene of interest

plot_single generates a barplot of a protein/gene of interest.

Usage

plot_single(object, proteins, type = c("contrast", "centered"), plot = TRUE)

Arguments

object

SummarizedExperiment, Data object for which differentially enriched proteins are annotated (output from test_diff() and add_rejections()).

proteins

Character, The name(s) of the protein(s)/gene(s) to plot.

type

'contrast' or 'centered', The type of data scaling used for plotting. Either the fold change ('contrast') or the centered log2-intensity ('centered').

plot

Logical(1), If TRUE (default) the barplot is produced. Otherwise (if FALSE), the summaries which the barplot is based on are returned.

Value

A barplot (generated by ggplot).

Examples

# Load example
data <- UbiLength
#> Error in eval(expr, envir, enclos): object 'UbiLength' not found
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
#> Error in data$Reverse: object of type 'closure' is not subsettable
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")
#> Error: proteins is not a data frame

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'grep': error in evaluating the argument 'x' in selecting a method for function 'colnames': object 'data_unique' not found
exp_design <- UbiLength_ExpDesign
#> Error in eval(expr, envir, enclos): object 'UbiLength_ExpDesign' not found
se <- make_se(data_unique, columns, exp_design)
#> Error in eval(expr, envir, enclos): object 'data_unique' not found

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
#> Error in filter_missval(se, thr = 0): could not find function "filter_missval"
norm <- normalize_vsn(filt)
#> Error in eval(expr, envir, enclos): object 'filt' not found
imputed <- impute(norm, fun = "MinProb", q = 0.01)
#> Error: se does not inherit from class SummarizedExperiment

# Test for differentially expressed proteins
diff <- test_diff(imputed, "control", "Ctrl")
#> Error in eval(expr, envir, enclos): object 'imputed' not found
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("standardGeneric", package = "methods")), new("nonstandardGenericFunction",     .Data = function (diff, alpha = 0.05, lfc = 1, thresholdmethod = "intersect",         curvature = 1, x0_fold = 2, ...)     {        standardGeneric("add_rejections")    }, generic = structure("add_rejections", package = "DEP2"),     package = "DEP2", group = list(), valueClass = character(0),     signature = c("diff", "alpha", "lfc", "thresholdmethod",     "curvature", "x0_fold"), default = NULL, skeleton = (function (diff,         alpha = 0.05, lfc = 1, thresholdmethod = "intersect",         curvature = 1, x0_fold = 2, ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "add_rejections"), domain = NA))(diff, alpha, lfc, thresholdmethod,         curvature, x0_fold, ...)), <environment>): unable to find an inherited method for function ‘add_rejections’ for signature ‘"standardGeneric"’

# Plot single proteins
plot_single(dep, 'USP15')
#> Error in eval(expr, envir, enclos): object 'dep' not found
plot_single(dep, 'USP15', 'centered')
#> Error in eval(expr, envir, enclos): object 'dep' not found
plot_single(dep, c('USP15', 'CUL1'))
#> Error in eval(expr, envir, enclos): object 'dep' not found
plot_single(dep, c('USP15', 'CUL1'), plot = FALSE)
#> Error in eval(expr, envir, enclos): object 'dep' not found