Differential expression test on a DESeqDataSet

test_diff_deg performs a differential expression test based on DESeq2 package and format the test results from DESeq() and results().

Usage

test_diff_deg(
  dds,
  type = c("all", "control", "manual"),
  control = NULL,
  test_contrasts = NULL,
  contrast_upon = "condition",
  filter_ihw = FALSE,
  independentFiltering = TRUE,
  lfcshark = FALSE,
  ...
)

Arguments

dds

A DESeqDataSet object.

type

"control", "all" or "manual", The type of contrasts that will be tested. This can be all possible pairwise comparisons ("all"), limited to the comparisons versus the control ("control"), or manually defined contrasts ("manual").

control

Character(1), The condition to which contrasts are generated if type = "control" (a control condition would be most appropriate).

test_contrasts

Character, The contrasts that will be tested if type = "manual". These should be formatted as "SampleA_vs_SampleB" or c("SampleA_vs_SampleC", "SampleB_vs_SampleC").

contrast_upon

Character, the contrast to bulid upon which column in experimentdesign.

filter_ihw

logical(1),Whether to use ihw to perform independent filtering and p-value adjustment, default FALSE. See results

independentFiltering

logical(1), whether independent filtering should be applied automatically, default TRUE. See results

lfcshark

logical(1),whether to use lfcshark, default FALSE. See results

...

Parameters transmitted to DESeq2