Generate a results table

get_results generates a results table from a proteomics or RNA-seq dataset on which differential analysis was performed.

Usage

get_results(object)

Arguments

object

SummarizedExperiment or DEGdata, (output from test_diff() and add_rejections()).

Value

A data.frame object containing all results variables from the performed analysis.

Examples

#Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Differential test
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
#> filter base on missing number is <= 0 in at least one condition.
#> filter base on missing number fraction < 0.4 in each row
#> filter base on giving formula 
norm <- normalize_vsn(filt)
#> vsn2: 8762 x 20 matrix (1 stratum). 
#> Please use 'meanSdPlot' to verify the fit.
imputed <- impute(norm, fun = "MinProb", q = 0.05)
#> [1] 0.3026531
diff <- test_diff(imputed, type = "control", control  = c("PBS"), fdr.type = "Storey's qvalue")
#> Tested contrasts: W10_vs_PBS, W2_vs_PBS, W4_vs_PBS, W6_vs_PBS, W9_vs_PBS
#> Storey's qvalue
dep <- add_rejections(diff, alpha = 0.01,lfc = 2)

# Get results
results <- get_results(dep)
colnames(results)
#>  [1] "name"                   "ID"                     "W10_vs_PBS_p.val"      
#>  [4] "W2_vs_PBS_p.val"        "W4_vs_PBS_p.val"        "W6_vs_PBS_p.val"       
#>  [7] "W9_vs_PBS_p.val"        "W10_vs_PBS_p.adj"       "W2_vs_PBS_p.adj"       
#> [10] "W4_vs_PBS_p.adj"        "W6_vs_PBS_p.adj"        "W9_vs_PBS_p.adj"       
#> [13] "W10_vs_PBS_significant" "W2_vs_PBS_significant"  "W4_vs_PBS_significant" 
#> [16] "W6_vs_PBS_significant"  "W9_vs_PBS_significant"  "significant"           
#> [19] "W10_vs_PBS_ratio"       "W2_vs_PBS_ratio"        "W4_vs_PBS_ratio"       
#> [22] "W6_vs_PBS_ratio"        "W9_vs_PBS_ratio"        "PBS_centered"          
#> [25] "W10_centered"           "W2_centered"            "W4_centered"           
#> [28] "W6_centered"            "W9_centered"           

significant_proteins <- results[results$significant,]
head(significant_proteins)
#>      name     ID W10_vs_PBS_p.val W2_vs_PBS_p.val W4_vs_PBS_p.val
#> 1    Acp5 Q05117     1.410945e-05    5.908297e-04    4.777297e-03
#> 2 Adamts8 P57110     1.506905e-01    2.794079e-02    1.354496e-02
#> 3    Add2 Q9QYB8     9.777005e-04    5.601892e-03    1.071916e-04
#> 4    Aif1 O70200     9.347648e-06    1.132573e-07    2.573544e-07
#> 5  Alox12 P39655     3.537141e-03    3.930261e-03    4.796279e-03
#> 6    Ank1 Q02357     2.100054e-04    4.825038e-04    1.664000e-05
#>   W6_vs_PBS_p.val W9_vs_PBS_p.val W10_vs_PBS_p.adj W2_vs_PBS_p.adj
#> 1    2.098882e-04    3.194426e-04         0.000675        2.47e-03
#> 2    5.281707e-04    6.672976e-03         0.142000        3.54e-02
#> 3    1.399617e-03    5.943734e-03         0.006460        1.13e-02
#> 4    1.146309e-05    8.108103e-04         0.000545        1.33e-05
#> 5    1.476214e-04    5.408129e-03         0.013100        8.76e-03
#> 6    7.308282e-05    7.948182e-05         0.002850        2.14e-03
#>   W4_vs_PBS_p.adj W6_vs_PBS_p.adj W9_vs_PBS_p.adj W10_vs_PBS_significant
#> 1        1.30e-02        0.001190         0.00277                   TRUE
#> 2        2.55e-02        0.002030         0.01670                  FALSE
#> 3        1.29e-03        0.003910         0.01550                   TRUE
#> 4        4.02e-05        0.000210         0.00465                   TRUE
#> 5        1.30e-02        0.000943         0.01450                  FALSE
#> 6        4.66e-04        0.000618         0.00128                   TRUE
#>   W2_vs_PBS_significant W4_vs_PBS_significant W6_vs_PBS_significant
#> 1                 FALSE                 FALSE                 FALSE
#> 2                 FALSE                 FALSE                  TRUE
#> 3                 FALSE                  TRUE                 FALSE
#> 4                  TRUE                  TRUE                 FALSE
#> 5                 FALSE                 FALSE                  TRUE
#> 6                 FALSE                  TRUE                  TRUE
#>   W9_vs_PBS_significant significant W10_vs_PBS_ratio W2_vs_PBS_ratio
#> 1                 FALSE        TRUE             2.26            1.29
#> 2                 FALSE        TRUE             1.42            1.86
#> 3                 FALSE        TRUE            -2.39           -1.56
#> 4                 FALSE        TRUE             2.48            2.84
#> 5                 FALSE        TRUE            -1.80           -1.45
#> 6                  TRUE        TRUE            -2.35           -1.76
#>   W4_vs_PBS_ratio W6_vs_PBS_ratio W9_vs_PBS_ratio PBS_centered W10_centered
#> 1           0.994            1.44            1.69        -1.14        1.120
#> 2           2.130            3.30            2.92        -1.89       -0.469
#> 3          -2.470           -1.87           -1.89         1.61       -0.786
#> 4           2.680            1.99            1.61        -1.91        0.567
#> 5          -1.410           -2.12           -1.70         1.35       -0.457
#> 6          -2.420           -2.12           -2.58         1.75       -0.593
#>   W2_centered W4_centered W6_centered W9_centered
#> 1     0.15100     -0.1470      0.3000       0.549
#> 2    -0.03530      0.2370      1.4100       1.030
#> 3     0.05130     -0.8600     -0.2650      -0.282
#> 4     0.93300      0.7680      0.0774      -0.299
#> 5    -0.10500     -0.0643     -0.7730      -0.351
#> 6    -0.00249     -0.6710     -0.3700      -0.826